+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - Change Grid Search base estimator param name to class
     - Remove Dataset cast to string preview
     - Add Error Analysis error standard deviation, drop midrange
+    - Naive Bayes Laplace smoothing no longer effects priors
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features